changed length type from uint16_t to size_t (ie uint32_t)

+++ Source/SerialBuffer.hpp
@@ -4,19 +4,19 @@
 #include <stdint.h>
 #include <string.h> // for memcpy
 
-template<uint16_t size, typename T = uint8_t>
+template<size_t size, typename T = uint8_t>
 class SerialBuffer {
 private:
-  volatile uint16_t writepos = 0;
-  volatile uint16_t readpos = 0;
+  volatile size_t writepos = 0;
+  volatile size_t readpos = 0;
   T buffer[size];
 public:
-  uint16_t getCapacity(){
+  size_t getCapacity(){
     return size;
   }
-  void push(T* data, uint16_t len){
+  void push(T* data, size_t len){
     T* dest = getWriteHead();
-    uint16_t rem = size-writepos;
+    size_t rem = size-writepos;
     if(len >= rem){
       memcpy(dest, data, rem);
       // note that len-rem may be zero
@@ -27,9 +27,9 @@ public:
       writepos += len;
     }
   }
-  void pull(T* data, uint16_t len){
+  void pull(T* data, size_t len){
     T* src = getReadHead();
-    uint16_t rem = size-readpos;
+    size_t rem = size-readpos;
     if(len >= rem){
       memcpy(data, src, rem);
       // note that len-rem may be zero
@@ -55,7 +55,7 @@ public:
 
   void skipUntilLast(char c){
     T* src = getReadHead();
-    uint16_t rem = size-readpos;
+    size_t rem = size-readpos;
     for(int i=0; i<rem; ++i){
       if(src[i] != c){
 	readpos += i;
@@ -74,7 +74,7 @@ public:
   T* getWriteHead(){
     return buffer+writepos;
   }
-  void incrementWriteHead(uint16_t len){
+  void incrementWriteHead(size_t len){
     // ASSERT((writepos >= readpos && writepos+len <= size) ||
     // 	   (writepos < readpos && writepos+len <= readpos), "uart rx overflow");
     writepos += len;
@@ -85,7 +85,7 @@ public:
   T* getReadHead(){
     return buffer+readpos;
   }
-  void incrementReadHead(uint16_t len){
+  void incrementReadHead(size_t len){
     // ASSERT((readpos >= writepos && readpos+len <= size) ||
     // 	   (readpos < writepos && readpos+len <= writepos), "uart rx underflow");
     readpos += len;
@@ -95,20 +95,20 @@ public:
   bool notEmpty(){
     return writepos != readpos;
   }
-  uint16_t getReadCapacity(){
+  size_t getReadCapacity(){
     return (writepos + size - readpos) % size;
   }
-  uint16_t getWriteCapacity(){
+  size_t getWriteCapacity(){
     return size - getReadCapacity();
   }
-  uint16_t getContiguousWriteCapacity(){
+  size_t getContiguousWriteCapacity(){
     if(writepos < readpos)
       return readpos - writepos;
     else
       return size - writepos;
     // return size-writepos;
   }
-  uint16_t getContiguousReadCapacity(){
+  size_t getContiguousReadCapacity(){
     if(writepos < readpos)
       return size - readpos;
     else
